requirements.txt
google_api_python_client==2.175.0
google_auth_oauthlib==1.2.2
openai==1.97.0
protobuf==6.31.1
python-dotenv==1.1.1
fastapi
uvicorn